Code cleanup & bug fix.
1) Test case logic embedded within dnvme could be removed. It was originally
   placed into the kernel in the days when the driver didn't support IOCTLS to
   reap and reap inquiry, thus special hooks wered needed during initial devel
2) Tabs where found were converted to 4 spaces.
3) Bugfix: An uint16_t was used to hold the value for the size, number of
   elements, of IOQ's. Since an IOQ can achieve a size of 65536, this value
   wasn't large enough to contain the neccessary value. Thus attempts to create
   the largest Q's results in undetermined behavior, however no kernel crashes
   were observed. Observations included CE's not arriving, and excessive delays
   during reaping. The repurcussions of this bug were found in tnvme, dnvme
   and in QEMU.

+++ dnvme_irq.c
@@ -1065,7 +1065,7 @@ exit:
  */
 int reap_inquiry_isr(struct metrics_cq  *pmetrics_cq_node,
     struct  metrics_device_list *pmetrics_device_elem,
-        u16 *num_remaining, u32 *isr_count)
+    u32 *num_remaining, u32 *isr_count)
 {
     u16 irq_no = pmetrics_cq_node->public_cq.irq_no; /* irq_no for CQ   */
     struct irq_track *pirq_node;
@@ -1089,12 +1089,13 @@ int reap_inquiry_isr(struct metrics_cq  *pmetrics_cq_node,
     /* Check if ISR is really fired for this CQ */
     if (pirq_node->isr_fired != 0) {
         /* process reap inquiry for isr fired case */
-        *num_remaining = reap_inquiry(pmetrics_cq_node, &pmetrics_device_elem
-            ->metrics_device->private_dev.pdev->dev);
+        *num_remaining = reap_inquiry(pmetrics_cq_node,
+            &pmetrics_device_elem->metrics_device->private_dev.pdev->dev);
     } else {
         /* To deal with ISR's aggregation, not supposed to notify CE's yet */
         *num_remaining = 0;
     }
+
     /* return the isr_count flag */
     *isr_count = pirq_node->isr_count;
     return SUCCESS;
@@ -1306,7 +1307,7 @@ int reset_isr_flag(struct metrics_device_list *pmetrics_device,
     struct irq_track *pirq_node; /* IRQ node inside irq track list */
     struct irq_cq_track *picq_node; /* CQ node inside irq node */
     struct metrics_cq  *pmetrics_cq_node; /* CQ node in metrics_cq_list */
-    u16 num_rem = 0;
+    u32 num_rem = 0;
 
     /* Get the Irq node for given irq vector */
     pirq_node = find_irq_node(pmetrics_device, irq_no);
